requirements.txt
plyfile
transforms3d.euler
tensorboardX
configparser
